Refactored Code - Added calendar_ui.py and calendar.kv file to view - Changed name of calendar.py file to calendar_ui.py - Remove old tkinker ui files - Calling calendar-ui in core.py to run view
