+++ AstroCal/view/__init__.py
@@ -1,2 +1,2 @@
 from view.console_output import *
-from view.tkGUI_Launch import *
+
