+++ AstroCal/view/calendar.kv
@@ -66,124 +66,124 @@
 			id: b4
 		Button:
 			id: b5
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b6
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b7
 	BoxLayout:
 		spacing: 10
 		Button:
 			id: b8
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b9
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b10
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b11
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b12
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b13
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b14
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 	BoxLayout:
 		spacing: 10
 		Button:
 			id: b15
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b16
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b17
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b18
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b19
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b20
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b21
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 	BoxLayout:
 		spacing: 10
 		Button:
 			id: b22
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b23
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b24
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b25
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b26
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b27
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b28
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 	BoxLayout:
 		spacing: 10
 		Button:
 			id: b29
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b30
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b31
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b32
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b33
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b34
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b35
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 	BoxLayout:
 		spacing: 10
 		Button:
 			id: b36
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b37
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b38
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b39
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b40
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b41
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
 		Button:
 			id: b42
-			on_press: app.showDayMode()
+			on_press: app.printToConsole()
